aider: Remove deprecated global seeding of random number generator and replace with local instance seeded with current time in nanoseconds in `generateLotteryNumbers` function.
# Aider chat conversation:

USER: fix this lint detection SA1019: rand.Seed has been deprecated since Go 1.20 and an alternative has been available since Go 1.0: As of Go 1.20 there is no reason to call Seed with a random value. Programs that call Seed with a known value to get a specific sequence of results should use New(NewSource(seed)) to obtain a local random generator.

ASSISTANT: To address the lint detection SA1019 regarding the deprecation of `rand.Seed` in Go 1.20, we need to modify the `cmd\lotto-numbers\generator.go` file. We will remove the `init` function that seeds the global random generator and instead create a local random generator within the `generateLotteryNumbers` function.

Here is the *SEARCH/REPLACE block* for the necessary changes:

cmd\lotto-numbers\generator.go
```go
<<<<<<< SEARCH
func init() {
	rand.Seed(time.Now().UnixNano())
}

=======
>>>>>>> REPLACE
```

cmd\lotto-numbers\generator.go
```go
<<<<<<< SEARCH
	for len(lotteryNumbers) < lines {
		rand.Shuffle(len(numbersList), func(i, j int) {
=======
	localRand := rand.New(rand.NewSource(time.Now().UnixNano()))
	for len(lotteryNumbers) < lines {
		localRand.Shuffle(len(numbersList), func(i, j int) {
>>>>>>> REPLACE
```

These changes will remove the deprecated global seeding of the random number generator and replace it with a local instance of a random generator that is seeded with the current time in nanoseconds. This ensures that the `generateLotteryNumbers` function uses a non-deprecated method of generating random numbers.