aider: Updated the listening port of the server to be a parameter that can be passed in at runtime.
# Aider chat conversation:

USER: update the listening port of the server to be a parameter that can be passed in at run time

ASSISTANT: To update the listening port of the server to be a parameter that can be passed at runtime, we need to modify `cmd\lotto-numbers\main.go` to:

1. Import the "os" package to read environment variables.
2. Retrieve the port number from an environment variable.
3. Use the retrieved port number when starting the server.

Here are the *SEARCH/REPLACE* blocks:

cmd\lotto-numbers\main.go
```go
<<<<<<< SEARCH
import (
	"log"
	"net/http"
=======
import (
	"log"
	"net/http"
	"os"
>>>>>>> REPLACE
```

cmd\lotto-numbers\main.go
```go
<<<<<<< SEARCH
	// Start the server
	log.Println("Starting server on :8080")
	if err := http.ListenAndServe(":8080", r); err != nil {
=======
	// Retrieve the port number from an environment variable or use 8080 as default
	port := os.Getenv("PORT")
	if port == "" {
		port = "8080"
	}

	// Start the server
	log.Printf("Starting server on :%s", port)
	if err := http.ListenAndServe(":"+port, r); err != nil {
>>>>>>> REPLACE
```

With these changes, you can now set the listening port by setting the `PORT` environment variable before running the application. If the `PORT` environment variable is not set, the server will default to listening on port 8080.